Renamed hasSize(int, int) to hasDimensions(int, int) for two-dimensional array asserts
