+++ src/main/java/org/assertj/core/internal/Int2DArrays.java
@@ -99,10 +99,10 @@ public class Int2DArrays {
    * @param expectedFirstDimensionSize the expected first dimension size of {@code actual}.
    * @param expectedSecondDimensionSize the expected second dimension size of {@code actual}.
    * @throws AssertionError if the given array is {@code null}.
-   * @throws AssertionError if the number of elements in the given array is different than the expected one.
+   * @throws AssertionError if the actual array's dimensions are not equal to the given ones.
    */
-  public void assertHasSize(AssertionInfo info, int[][] actual, int expectedFirstDimensionSize,
-    int expectedSecondDimensionSize) {
+  public void assertHasDimensions(AssertionInfo info, int[][] actual, int expectedFirstDimensionSize,
+                                  int expectedSecondDimensionSize) {
     arrays.assertHasSize(info, actual, expectedFirstDimensionSize);
     for (int i = 0; i < actual.length; i++) {
       arrays.assertHasSize(info, failures, actual[i], expectedSecondDimensionSize, i);
