+++ src/test/java/org/assertj/core/internal/int2darrays/Int2DArrays_assertHasDimensions_Test.java
@@ -10,7 +10,7 @@
  *
  * Copyright 2012-2020 the original author or authors.
  */
-package org.assertj.core.internal.char2darrays;
+package org.assertj.core.internal.int2darrays;
 
 import static org.assertj.core.api.BDDAssertions.then;
 import static org.assertj.core.error.ShouldHaveSize.shouldHaveSize;
@@ -19,23 +19,23 @@ import static org.assertj.core.test.TestData.someInfo;
 import static org.assertj.core.util.AssertionsUtil.expectAssertionError;
 
 import org.assertj.core.api.AssertionInfo;
-import org.assertj.core.internal.Char2DArrays;
-import org.assertj.core.internal.Char2DArraysBaseTest;
+import org.assertj.core.internal.Int2DArrays;
+import org.assertj.core.internal.Int2DArraysBaseTest;
 import org.junit.jupiter.api.Test;
 
 /**
- * Tests for <code>{@link Char2DArrays#assertHasSize(AssertionInfo, char[][], int, int)}</code>.
+ * Tests for <code>{@link Int2DArrays#assertHasDimensions(AssertionInfo, int[][], int, int)}</code>.
  *
  * @author Maciej Wajcht
  */
-public class Char2DArrays_assertHasSize_Test extends Char2DArraysBaseTest {
+public class Int2DArrays_assertHasDimensions_Test extends Int2DArraysBaseTest {
 
   @Test
   public void should_fail_if_actual_is_null() {
     // GIVEN
-    char[][] actual = null;
+    int[][] actual = null;
     // WHEN
-    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasSize(someInfo(), actual, 2, 3));
+    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasDimensions(someInfo(), actual, 2, 3));
     // THEN
     then(assertionError).hasMessage(shouldNotBeNull().create());
   }
@@ -45,7 +45,7 @@ public class Char2DArrays_assertHasSize_Test extends Char2DArraysBaseTest {
     // GIVEN
     int expectedFirstDimensionSize = 10;
     // WHEN
-    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasSize(someInfo(), actual,
+    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasDimensions(someInfo(), actual,
                                                                                     expectedFirstDimensionSize, 3));
     // THEN
     then(assertionError).hasMessage(shouldHaveSize(actual, actual.length, expectedFirstDimensionSize).create());
@@ -56,14 +56,14 @@ public class Char2DArrays_assertHasSize_Test extends Char2DArraysBaseTest {
     // GIVEN
     int expectedSecondDimensionSize = 10;
     // WHEN
-    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasSize(someInfo(), actual, 2,
+    AssertionError assertionError = expectAssertionError(() -> arrays.assertHasDimensions(someInfo(), actual, 2,
                                                                                     expectedSecondDimensionSize));
     // THEN
-    then(assertionError).hasMessage(shouldHaveSize(new char[] { 'a', 'b', 'c' }, 3, expectedSecondDimensionSize, 0).create());
+    then(assertionError).hasMessage(shouldHaveSize(new int[] { 0, 2, 4 }, 3, expectedSecondDimensionSize, 0).create());
   }
 
   @Test
   public void should_pass_if_size_of_actual_is_equal_to_expected_size() {
-    arrays.assertHasSize(someInfo(), actual, 2, 3);
+    arrays.assertHasDimensions(someInfo(), actual, 2, 3);
   }
 }
